corrigi a indentação e espaços dos exercícios e novamente testados a compilação. do 1062 ao 1180c.c

+++ 1176.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    long long f[61];
+    double f[61];
     int t, n, i;
 
     f[0] = 0;
@@ -10,10 +10,10 @@ int main() {
     for (i = 2; i <= 60; i++)
         f[i] = f[i-1] + f[i-2];
 
-    scanf("%d", &t);
+    scanf ("%d", &t);
     while (t--) {
-        scanf("%d", &n);
-        printf("Fib(%d) = %lld\n", n, f[n]);
+        scanf ("%d", &n);
+        printf ("Fib(%d) = %.0f\n", n, f[n]);
     }
 
     return 0;
